refactor: context组件优化
